Moved extension-functions to separate files
