+++ src/main/kotlin/dev/psygamer/wireframe/block/entity/TickingBlockEntity.kt
@@ -2,7 +2,7 @@ package dev.psygamer.wireframe.block.entity
 
 import dev.psygamer.wireframe.util.Identifier
 
-open class TickingBlockEntity constructor(identifier: Identifier) : BlockEntity(identifier) {
+open class TickingBlockEntity protected constructor(identifier: Identifier) : BlockEntity(identifier) {
 	
 	open fun tick() {}
 }
