+++ src/main/kotlin/dev/psygamer/wireframe/block/state/BlockStateDefinition.kt
@@ -2,7 +2,6 @@ package dev.psygamer.wireframe.block.state
 
 import dev.psygamer.wireframe.block.Block
 import dev.psygamer.wireframe.block.state.property.BlockProperty
-import dev.psygamer.wireframe.internal.block.NativeBlockStateDefinition
 
 interface BlockStateDefinition {
 	
@@ -14,6 +13,3 @@ interface BlockStateDefinition {
 	
 	fun getProperty(propertyName: String): BlockProperty<*>?
 }
-
-internal val net.minecraft.state.StateContainer<net.minecraft.block.Block, net.minecraft.block.BlockState>.wfWrapped: BlockStateDefinition
-	get() = NativeBlockStateDefinition(this)
