+++ src/main/kotlin/dev/psygamer/wireframe/util/BlockPosition.kt
@@ -21,9 +21,3 @@ class BlockPosition(x: Int, y: Int, z: Int) : Vector3i(x, y, z) {
 		)
 	}
 }
-
-internal val net.minecraft.util.math.BlockPos.wfWrapped: BlockPosition
-	get() = BlockPosition(this.x, this.y, this.z)
-
-internal val BlockPosition.mcNative: net.minecraft.util.math.BlockPos
-	get() = net.minecraft.util.math.BlockPos(this.x, this.y, this.z)
