+++ src/main/kotlin/dev/psygamer/wireframe/util/Identifier.kt
@@ -9,12 +9,3 @@ data class Identifier @JvmOverloads constructor(
 		return "$namespace:$path"
 	}
 }
-
-internal val net.minecraft.util.ResourceLocation.wfWrapped: Identifier
-	get() {
-		return Identifier(this.namespace, this.path)
-	}
-internal val Identifier.mcNative: net.minecraft.util.ResourceLocation
-	get() {
-		return net.minecraft.util.ResourceLocation(this.namespace, this.path)
-	}
