+++ src/main/kotlin/dev/psygamer/wireframe/util/math/vector/Vector2f.kt
@@ -5,6 +5,7 @@ import kotlin.math.atan2
 import kotlin.math.sqrt
 
 open class Vector2f(val x: Float, val y: Float) {
+	
 	operator fun plus(other: Vector2f): Vector2f {
 		return Vector2f(this.x + other.x, this.y + other.y)
 	}
@@ -61,10 +62,3 @@ open class Vector2f(val x: Float, val y: Float) {
 		return acos(this dot other / (this.magnitude * other.magnitude))
 	}
 }
-
-val net.minecraft.util.math.vector.Vector2f.wfWrapped: Vector2f
-	get() = Vector2f(this.x, this.y)
-
-val Vector2f.mcNative: net.minecraft.util.math.vector.Vector2f
-	get() = net.minecraft.util.math.vector.Vector2f(this.x, this.y)
-
